+++ src/internal/m365/graph/collections_test.go
@@ -24,10 +24,10 @@ func (suite *CollectionsUnitSuite) TestNewPrefixCollection() {
 	serv := path.OneDriveService
 	cat := path.FilesCategory
 
-	p1, err := path.ServicePrefix("t", "ro1", serv, cat)
+	p1, err := path.BuildPrefix("t", "ro1", serv, cat)
 	require.NoError(t, err, clues.ToCore(err))
 
-	p2, err := path.ServicePrefix("t", "ro2", serv, cat)
+	p2, err := path.BuildPrefix("t", "ro2", serv, cat)
 	require.NoError(t, err, clues.ToCore(err))
 
 	items, err := path.Build("t", "ro", serv, cat, true, "fld", "itm")
